error handled and login logic modified
