--- src/app/(auth)/login/page.css
@@ -1,68 +1,0 @@
-@import "@radix-ui/colors/red.css";
-@import "@radix-ui/colors/green.css";
-
-.Alerticon,
-.SuccessIcon {
-  width: 1.2rem;
-  height: 1.2rem;
-}
-html,
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  overflow-x: hidden;
-  max-height: 100%;
-  overflow-y: hidden;
-}
-/* 전체 컨테이너: 화면 중앙 정렬 + 배경 */
-#LoginContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  background-color: #f9f9f9;
-}
-
-/* 로그인 폼 카드 스타일 */
-#LoginForm {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  width: 100%;
-  max-width: 26rem;
-  padding: 3rem 2rem;
-  background-color: white;
-  border-radius: 1rem;
-  box-shadow: 0 6px 24px rgba(0, 0, 0, 0.08);
-  gap: 1.5rem;
-  text-align: center;
-}
-
-/* Google 로그인 버튼 wrapper */
-#GoogleLoginWrapper {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-/* 구글 버튼 자체 스타일 (iframe 안이라 제한적) */
-.g_id_signin {
-  width: 100% !important;
-  max-width: 20rem !important; /* 20rem ≈ 320px */
-  display: flex !important;
-  justify-content: center !important;
-}
-
-/* 회원가입 버튼 컨테이너 */
-#LoginButtonsContainer {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-}
-
-/* 버튼 자체 */
-#LoginButtonsContainer button {
-  width: 100%;
-  max-width: 20rem;
-}
