+++ src/app/about/rules/page.css
