+++ src/app/about/rules/page.jsx
