+++ src/app/board/[id]/page.css
@@ -8,14 +8,14 @@
   align-items: center;
   max-width: 60rem;
   margin: 0 auto;
-  padding: 2rem 1rem;
+  padding: 1.5rem 1rem;
 }
 
 #SigList {
   width: 100%;
   display: flex;
   flex-direction: column;
-  gap: 1.25rem;
+  gap: 0.75rem;
 }
 
 .sigLink {
@@ -25,17 +25,17 @@
 
 .sigCard {
   border: 1px solid var(--slate-6);
-  border-radius: 0.75rem;
+  border-radius: 0.5rem;
   background: white;
-  padding: 1rem 1.25rem;
+  padding: 0.75rem 1rem;
   display: flex;
   flex-direction: column;
-  gap: 0.4rem;
-  transition: box-shadow 0.2s ease;
+  gap: 0.3rem;
+  transition: background-color 0.2s ease;
 }
 
 .sigCard:hover {
-  box-shadow: 0 2px 6px var(--slate-5);
+  background-color: var(--slate-2);
 }
 
 .sigTopbar {
@@ -45,20 +45,20 @@
 }
 
 .sigTitle {
-  font-size: 1.1rem;
+  font-size: 1rem;
   font-weight: 600;
   color: var(--indigo-11);
 }
 
 .sigUserCount {
-  font-size: 0.9rem;
-  color: var(--slate-10);
+  font-size: 0.85rem;
+  color: var(--slate-9);
 }
 
 .sigDescription {
-  font-size: 0.95rem;
-  color: var(--slate-12);
-  line-height: 1.4;
+  font-size: 0.9rem;
+  color: var(--slate-11);
+  line-height: 1.3;
   white-space: pre-wrap;
 }
 
@@ -82,9 +82,9 @@
 
 .SigHeader {
   width: 100%;
-  max-width: 60rem; /* SigListContainer와 동일하게 제한 */
+  max-width: 60rem;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  margin-bottom: 2rem;
+  margin-bottom: 1.25rem;
 }
