+++ src/app/footer.css
@@ -1,8 +1,6 @@
 @import "@radix-ui/colors/slate.css";
 
 #Footer {
-  position: relative;
-  bottom: 0px;
   display: flex;
   justify-content: space-between;
   align-items: flex-start;
