+++ src/app/sig/create/CreateSigClient.jsx
@@ -23,8 +23,7 @@ export default function CreateSigClient() {
   useEffect(() => {
     const jwt = localStorage.getItem("jwt");
     if (!jwt) {
-      alert("로그인이 필요합니다.");
-      router.push("/login");
+      router.push("/us/login");
     }
   }, [router]);
 
@@ -37,8 +36,7 @@ export default function CreateSigClient() {
   const onSubmit = async (data) => {
     const jwt = localStorage.getItem("jwt");
     if (!jwt) {
-      alert("로그인이 필요합니다.");
-      router.push("/login");
+      router.push("/us/login");
       return;
     }
